Rename subpackages and modules
Renamed subpackages and modules so that they represent their contents more accurately, which is required for future implementations.
