+++ src/fs_mapping_tools/__init__.py
@@ -1,3 +1,3 @@
 from .track.cones import Cone, ConeArray
-from .vehicle.camera import Camera
+from .vehicle.detection import Camera
 from .vehicle.car import Car, CarState, CarStructure
