+++ tests/modular/vehicle/detection/interactive_sim.json
